🔧  Use plop to generate new components
Generates new component scaffolds, along with tests and indexes in
the parent directory from the command 'plop'

### Why Plop?

1. It's quick. Four characters
2. It's memorable. You already can't get it out of your head
3. It's fun.

Think of a magical rabbit running around and plopping react components.
you can't get this out of your head now. It's going to haunt every
second of your existence from birthdays to bar mitzvahs
to your child's high school graduation. You will always remember the
significance of the term "plop" as it related to front-end dev work.

### Okay. What about the __template__ dir?

I like the idea of having the templates for component config
in the same location as the components. Like things together.